Add code comments
I want the code to be as comprehensible as possible, so I am adding some
comments to help guide the reader along.

While I was at it, I ended up nesting a bit of the breakpoint-width
function under an else conditional, since it was effectively an else
anyway, but I thought it would be more readable this way.

Change-Id: I9cfb6a0b7d15ff63e42ce5e03ff456211ff03a37
Reviewed-on: https://gerrit.causes.com/31488
Reviewed-by: Joe Lencioni <joe.lencioni@causes.com>
Tested-by: Joe Lencioni <joe.lencioni@causes.com>